+++ packages/varlet-ui/types/icon.d.ts
@@ -1,12 +1,12 @@
-import { VarComponent } from './varComponent'
+import { BasicAttributes, VarComponent } from './varComponent'
 
-export interface IconProps {
+export interface IconProps extends BasicAttributes {
   name?: string
   size?: string | number
   color?: string
   namespace?: string
   transition?: string | number
-  onClick?: () => (event: Event) => void
+  onClick?: (event: Event) => void
 }
 
 export class Icon extends VarComponent {