+++ scripts/build.mjs
@@ -6,8 +6,11 @@ const CWD = process.cwd()
 const PKG_CLI = resolve(CWD, './packages/varlet-cli')
 const PKG_ICONS = resolve(CWD, './packages/varlet-icons')
 const PKG_UI = resolve(CWD, './packages/varlet-ui')
+const PKG_SHARED = resolve(CWD, './packages/varlet-shared')
 
 export const buildCli = () => execa('pnpm', ['build'], { cwd: PKG_CLI })
+
+export const buildShared = () => execa('pnpm', ['build'], { cwd: PKG_SHARED })
 
 export const buildIcons = () => execa('pnpm', ['build'], { cwd: PKG_ICONS })